fix a compile warning
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@5681 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/FreeSerialPort/FreeSerialPort.cpp
@@ -236,7 +236,7 @@ int FreeSerialPort::OnCommunicationSetting(MM::PropertyBase* pProp, MM::ActionTy
       // get the serial device
       MM::Device* pS = GetCoreCallback()->GetDevice(this, port_.c_str());
       pS->Shutdown();
-      size_t np = communicationSettings_.size();
+      int np =  static_cast<int>(communicationSettings_.size());
       for(int ip = 0; ip < np; ++ip)
       {
         if( indexx == ip)
